refactor: Convert community course stage solution comment model to typescript
